pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "llm-wiki"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    # API
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "python-multipart>=0.0.9",
    "sse-starlette>=1.8.2",

    # DB
    "sqlalchemy[asyncio]>=2.0.30",
    "asyncpg>=0.29.0",
    "alembic>=1.13.1",
    "pgvector>=0.2.5",

    # Auth
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt<4.0.0",

    # LLM
    "anthropic>=0.28.0",
    "voyageai>=0.2.3",
    "pymupdf>=1.24.0",

    # Workers
    "celery[redis]>=5.4.0",
    "redis[hiredis]>=5.0.4",

    # Storage
    "boto3>=1.34.0",
    "aiofiles>=23.2.1",

    # Git
    "gitpython>=3.1.43",

    # Graph
    "networkx>=3.3",

    # Utils
    "pydantic[email]>=2.7.0",
    "pydantic-settings>=2.3.0",
    "python-dotenv>=1.0.1",
    "structlog>=24.2.0",
    "tenacity>=8.3.0",
    "httpx>=0.27.0",
    "tiktoken>=0.7.0",
    "sse-starlette>=2.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.7",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
    "factory-boy>=3.3.0",
    "ruff>=0.4.7",
    "mypy>=1.10.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
